QR Gen/requirements.txt
pyqrcode
Pillow
tk
